@@ -1,13 +1,13 @@
 [build-system]
-requires = ["setuptools>=68", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [project]
 name = "pinescript-syntax-checker"
-version = "0.1.0"
+version = "0.2.0"
 description = "PineScript syntax checker MCP server and CLI"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.13"
 license = { file = "LICENSE" }
 authors = [{ name = "erevus" }]
 dependencies = [
@@ -20,4 +20,52 @@ Homepage = "https://github.com/erevus-cn/pinescript-syntax-checker"
 
 [project.scripts]
 pinescript-syntax-checker = "pinescript_syntax_checker.server:main"
-pinescript-syntax-checker-cli = "pinescript_syntax_checker.pinescript_checker:main"
+pinescript-checker = "pinescript_syntax_checker.pinescript_checker:main"
+
+[dependency-groups]
+dev = [
+  "ruff>=0.8.0",
+  "ty>=0.0.1",
+  "pre-commit>=3.0.0",
+]
+
+[tool.ruff]
+line-length = 100
+target-version = "py313"
+fix = true
+
+[tool.ruff.lint]
+select = [
+  "E",     # pycodestyle errors
+  "W",     # pycodestyle warnings
+  "F",     # pyflakes
+  "I",     # isort
+  "N",     # pep8-naming
+  "UP",    # pyupgrade
+  "ANN",   # flake8-annotations
+  "ASYNC", # flake8-async
+  "B",     # flake8-bugbear
+  "C4",    # flake8-comprehensions
+  "DTZ",   # flake8-datetimez
+  "T10",   # flake8-debugger
+  "FA",    # flake8-future-annotations
+  "ISC",   # flake8-implicit-str-concat
+  "PIE",   # flake8-pie
+  "PT",    # flake8-pytest-style
+  "RSE",   # flake8-raise
+  "RET",   # flake8-return
+  "SIM",   # flake8-simplify
+  "TCH",   # flake8-type-checking
+  "ARG",   # flake8-unused-arguments
+  "PTH",   # flake8-use-pathlib
+  "PL",    # pylint
+  "RUF",   # ruff-specific rules
+]
+ignore = []
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401"]
+
+[tool.ty]
+# ty configuration - currently supports environment, src, rules, terminal, overrides
+# See: https://github.com/astral-sh/ty for configuration options
